# frontend/packages/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "需脉枢纽 API",
    "description": "需脉枢纽 · B2B 代工制造供需智能语义匹配平台",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/vendor/register": {
      "post": {
        "tags": [
          "vendor"
        ],
        "summary": "企业基本信息（厂商注册详情）",
        "operationId": "register_vendor_api_v1_vendor_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_VendorOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/vendor/{vendor_id}": {
      "get": {
        "tags": [
          "vendor"
        ],
        "summary": "厂商档案",
        "operationId": "get_vendor_api_v1_vendor__vendor_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vendor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vendor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_VendorOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vendor/capability/upload": {
      "post": {
        "tags": [
          "vendor"
        ],
        "summary": "能力录入（仅上传文档，AI 解析）",
        "operationId": "upload_capability_api_v1_vendor_capability_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_capability_api_v1_vendor_capability_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CapabilityOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vendor/capability/{vendor_id}": {
      "get": {
        "tags": [
          "vendor"
        ],
        "summary": "厂商能力档案（只读）",
        "operationId": "get_capability_api_v1_vendor_capability__vendor_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vendor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vendor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CapabilityOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vendor/capability/{vendor_id}/documents/{document_id}": {
      "delete": {
        "tags": [
          "vendor"
        ],
        "summary": "删除能力文档（触发重新解析，版本+1）",
        "operationId": "delete_capability_document_api_v1_vendor_capability__vendor_id__documents__document_id__delete",
        "parameters": [
          {
            "name": "vendor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vendor Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CapabilityOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_CapabilityOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CapabilityOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[CapabilityOut]"
      },
...
      "ApiResponse_VendorOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VendorOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[VendorOut]"
      },
...
      "Body_upload_capability_api_v1_vendor_capability_upload_post": {
        "properties": {
          "vendor_id": {
            "type": "string",
            "title": "Vendor Id"
          },
          "documents": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Documents",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "vendor_id"
        ],
        "title": "Body_upload_capability_api_v1_vendor_capability_upload_post"
      },
...
      "CapabilityOut": {
        "properties": {
          "capability_id": {
            "type": "string",
            "title": "Capability Id"
          },
          "vendor_id": {
            "type": "string",
            "title": "Vendor Id"
          },
          "structured_tags": {
            "type": "object",
            "title": "Structured Tags"
          },
          "summary_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary Text"
          },
          "audit_status": {
            "type": "string",
            "enum": [
              "pending",
              "passed",
              "rejected"
            ],
            "title": "Audit Status",
            "default": "pending"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "description": "档案版本号：每次重新解析（增/删文档）递增",
            "default": 1
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "档案生成时间"
          },
          "doc_count": {
            "type": "integer",
            "title": "Doc Count",
            "description": "本次档案基于的文档数",
            "default": 0
          },
          "completeness": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completeness",
            "description": "能力完备度：硬能力已提取字段/硬能力总数 0-1"
          },
          "source_map": {
            "type": "object",
            "title": "Source Map",
            "description": "能力字段 → 文档来源（含字段级解析置信度 confidence 0-1，<0.6 低置信度）"
          },
          "raw_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Text",
            "description": "文档原文片段"
          },
          "doc_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Doc Urls",
            "description": "文档名列表"
          }
        },
        "type": "object",
        "required": [
          "capability_id",
          "vendor_id"
        ],
        "title": "CapabilityOut",
        "description": "厂商能力档案（只读，01D）：AI 从文档解析，硬/软能力分层，全部可溯源到文档。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VendorOut": {
        "properties": {
          "vendor_id": {
            "type": "string",
            "title": "Vendor Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "main_industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Main Industry"
          },
          "credit_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credit Code"
          },
          "audit_status": {
            "type": "string",
            "title": "Audit Status",
            "default": "pending"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "vendor_id",
          "company_name",
          "created_at"
        ],
        "title": "VendorOut"
      },
      "VendorRegisterRequest": {
        "properties": {
          "company_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Company Name"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "main_industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Main Industry"
          },
          "credit_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credit Code",
            "description": "统一社会信用代码（全局唯一）"
          },
          "license_file_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License File Id",
            "description": "营业执照文件 ID（先调 /files/upload）"
          }
        },
        "type": "object",
        "required": [
          "company_name"
        ],
        "title": "VendorRegisterRequest",
        "description": "企业基本信息（01A 厂商注册详情 / 产品 1.1）。"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}